Fixes #MR-634: Trennung von FileResource-Metadaten & -Upload
