+++ dc-cudami-admin/dc-cudami-admin-backend-api/src/main/java/de/digitalcollections/cudami/admin/backend/api/repository/identifiable/resource/FileResourceMetadataRepository.java
@@ -1,0 +1,6 @@
+package de.digitalcollections.cudami.admin.backend.api.repository.identifiable.resource;
+
+import de.digitalcollections.cudami.admin.backend.api.repository.identifiable.IdentifiableRepository;
+import de.digitalcollections.model.api.identifiable.resource.FileResource;
+
+public interface FileResourceMetadataRepository extends IdentifiableRepository<FileResource> {}
