+++ dc-cudami-admin/dc-cudami-admin-backend-rest/pom.xml
@@ -44,10 +44,6 @@
       <artifactId>feign-jackson</artifactId>
     </dependency>
     <dependency>
-      <groupId>io.github.openfeign.form</groupId>
-      <artifactId>feign-form-spring</artifactId>
-    </dependency>
-    <dependency>
       <groupId>org.apache.commons</groupId>
       <artifactId>commons-lang3</artifactId>
     </dependency>
