+++ dc-cudami-admin/dc-cudami-admin-backend-rest/src/main/java/de/digitalcollections/cudami/admin/backend/impl/repository/identifiable/resource/FileResourceBinaryRepositoryImpl.java
@@ -1,21 +1,13 @@
 package de.digitalcollections.cudami.admin.backend.impl.repository.identifiable.resource;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import de.digitalcollections.cudami.admin.backend.api.repository.identifiable.resource.CudamiFileResourceRepository;
-import de.digitalcollections.cudami.admin.backend.impl.repository.identifiable.IdentifiableRepositoryImpl;
-import de.digitalcollections.cudami.admin.backend.impl.repository.identifiable.IdentifiableRepositoryImpl.FindParams;
-import de.digitalcollections.model.api.http.exceptions.client.ResourceNotFoundException;
+import de.digitalcollections.cudami.admin.backend.api.repository.identifiable.resource.FileResourceBinaryRepository;
 import de.digitalcollections.model.api.identifiable.resource.FileResource;
 import de.digitalcollections.model.api.identifiable.resource.exceptions.ResourceIOException;
-import de.digitalcollections.model.api.paging.PageRequest;
-import de.digitalcollections.model.api.paging.PageResponse;
-import de.digitalcollections.model.impl.identifiable.resource.FileResourceImpl;
-import feign.form.FormData;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
-import java.util.UUID;
 import org.apache.http.HttpEntity;
 import org.apache.http.HttpResponse;
 import org.apache.http.client.HttpClient;
@@ -31,73 +23,15 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class CudamiFileResourceRepositoryImpl extends IdentifiableRepositoryImpl<FileResource>
-    implements CudamiFileResourceRepository {
+public class FileResourceBinaryRepositoryImpl implements FileResourceBinaryRepository {
 
   private static final Logger LOGGER =
-      LoggerFactory.getLogger(CudamiFileResourceRepositoryImpl.class);
+      LoggerFactory.getLogger(FileResourceBinaryRepositoryImpl.class);
 
   @Value(value = "${cudami.server.address}")
   private String cudamiServerAddress;
 
-  @Autowired private CudamiFileResourceRepositoryEndpoint endpoint;
-
   @Autowired ObjectMapper objectMapper;
-
-  @Override
-  public long count() {
-    return endpoint.count();
-  }
-
-  @Override
-  public FileResource create() {
-    return new FileResourceImpl();
-  }
-
-  @Override
-  public PageResponse<FileResource> find(PageRequest pageRequest) {
-    FindParams f = getFindParams(pageRequest);
-    PageResponse<FileResource> pageResponse =
-        endpoint.find(
-            f.getPageNumber(),
-            f.getPageSize(),
-            f.getSortField(),
-            f.getSortDirection(),
-            f.getNullHandling());
-    return getGenericPageResponse(pageResponse);
-  }
-
-  @Override
-  public FileResource findOneByIdentifier(String namespace, String id) {
-    try {
-      return endpoint.findOneByIdentifier(namespace, id);
-    } catch (ResourceNotFoundException e) {
-      return null;
-    }
-  }
-
-  @Override
-  public FileResource findOne(UUID uuid) {
-    return endpoint.findOne(uuid);
-  }
-
-  @Override
-  public FileResource save(FileResource fileResource) {
-    return endpoint.save(fileResource);
-  }
-
-  @Override
-  public FileResource save(FileResource fileResource, byte[] bytes) {
-    String contentType = fileResource.getMimeType().getTypeName();
-    String fileName = fileResource.getFilename();
-    FormData formData = new FormData(contentType, fileName, bytes);
-    return endpoint.save(fileResource, formData);
-  }
-
-  @Override
-  public FileResource update(FileResource fileResource) {
-    return endpoint.update(fileResource.getUuid(), fileResource);
-  }
 
   @Override
   public FileResource upload(InputStream inputStream, String filename, String contentType)
@@ -137,7 +71,7 @@ public class CudamiFileResourceRepositoryImpl extends IdentifiableRepositoryImpl
 
   private FileResource doPost(HttpEntity entity)
       throws UnsupportedOperationException, IOException, ResourceIOException {
-    HttpPost post = new HttpPost(cudamiServerAddress + "/latest/fileresources/new/upload");
+    HttpPost post = new HttpPost(cudamiServerAddress + "/latest/files");
     post.setEntity(entity);
     HttpClient client = HttpClientBuilder.create().build();
     HttpResponse response = client.execute(post);
