+++ dc-cudami-admin/dc-cudami-admin-backend-rest/src/main/java/de/digitalcollections/cudami/admin/backend/impl/repository/identifiable/resource/FileResourceMetadataRepositoryEndpoint.java
@@ -6,10 +6,9 @@ import de.digitalcollections.model.api.paging.PageResponse;
 import feign.Headers;
 import feign.Param;
 import feign.RequestLine;
-import feign.form.FormData;
 import java.util.UUID;
 
-public interface CudamiFileResourceRepositoryEndpoint extends RepositoryEndpoint {
+public interface FileResourceMetadataRepositoryEndpoint extends RepositoryEndpoint {
 
   @RequestLine("GET /latest/fileresources/count")
   long count();
@@ -30,20 +29,9 @@ public interface CudamiFileResourceRepositoryEndpoint extends RepositoryEndpoint
   @Headers("Accept: application/json")
   FileResource findOneByIdentifier(@Param("namespace") String namespace, @Param("id") String id);
 
-  /**
-   * only saving non binary parts of resource
-   *
-   * @param fileresource metadata object
-   * @return saved object
-   */
-  @RequestLine("POST /latest/fileresources2") // FIXME conflict of two post methods with same url...
+  @RequestLine("POST /latest/fileresources")
   @Headers("Content-Type: application/json")
   FileResource save(FileResource fileresource);
-
-  @RequestLine("POST /latest/fileresources")
-  @Headers("Content-Type: multipart/form-data")
-  FileResource save(
-      @Param("fileresource") FileResource fileresource, @Param("binaryData") FormData binaryData);
 
   @RequestLine("PUT /latest/fileresources/{uuid}")
   @Headers("Content-Type: application/json")
