+++ dc-cudami-admin/dc-cudami-admin-business/src/main/java/de/digitalcollections/cudami/admin/business/api/service/identifiable/resource/FileResourceBinaryService.java
@@ -1,14 +1,10 @@
 package de.digitalcollections.cudami.admin.business.api.service.identifiable.resource;
 
-import de.digitalcollections.cudami.admin.business.api.service.exceptions.IdentifiableServiceException;
-import de.digitalcollections.cudami.admin.business.api.service.identifiable.IdentifiableService;
 import de.digitalcollections.model.api.identifiable.resource.FileResource;
 import de.digitalcollections.model.api.identifiable.resource.exceptions.ResourceIOException;
 import java.io.InputStream;
 
-public interface CudamiFileResourceService extends IdentifiableService<FileResource> {
-
-  FileResource save(FileResource fileResource, byte[] bytes) throws IdentifiableServiceException;
+public interface FileResourceBinaryService {
 
   public FileResource upload(InputStream inputStream, String filename, String contentType)
       throws ResourceIOException;
