+++ dc-cudami-admin/dc-cudami-admin-business/src/main/java/de/digitalcollections/cudami/admin/business/api/service/identifiable/resource/FileResourceMetadataService.java
@@ -1,0 +1,6 @@
+package de.digitalcollections.cudami.admin.business.api.service.identifiable.resource;
+
+import de.digitalcollections.cudami.admin.business.api.service.identifiable.IdentifiableService;
+import de.digitalcollections.model.api.identifiable.resource.FileResource;
+
+public interface FileResourceMetadataService extends IdentifiableService<FileResource> {}
