+++ dc-cudami-server/dc-cudami-server-business/src/main/java/de/digitalcollections/cudami/server/business/impl/service/identifiable/resource/FileResourceBinaryServiceImpl.java
@@ -3,42 +3,29 @@ package de.digitalcollections.cudami.server.business.impl.service.identifiable.r
 import de.digitalcollections.cudami.server.backend.api.repository.identifiable.resource.FileResourceBinaryRepository;
 import de.digitalcollections.cudami.server.backend.api.repository.identifiable.resource.FileResourceMetadataRepository;
 import de.digitalcollections.cudami.server.business.api.service.exceptions.IdentifiableServiceException;
-import de.digitalcollections.cudami.server.business.api.service.identifiable.resource.FileResourceService;
-import de.digitalcollections.cudami.server.business.impl.service.identifiable.IdentifiableServiceImpl;
+import de.digitalcollections.cudami.server.business.api.service.identifiable.resource.FileResourceBinaryService;
 import de.digitalcollections.model.api.identifiable.resource.FileResource;
-import de.digitalcollections.model.api.identifiable.resource.MimeType;
 import de.digitalcollections.model.api.identifiable.resource.exceptions.ResourceIOException;
 import java.io.InputStream;
-import java.util.UUID;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
-public class FileResourceServiceImpl extends IdentifiableServiceImpl<FileResource>
-    implements FileResourceService {
+public class FileResourceBinaryServiceImpl implements FileResourceBinaryService {
 
-  private static final Logger LOGGER = LoggerFactory.getLogger(FileResourceServiceImpl.class);
+  private static final Logger LOGGER = LoggerFactory.getLogger(FileResourceBinaryServiceImpl.class);
 
   private final FileResourceBinaryRepository binaryRepository;
+  private final FileResourceMetadataRepository metadataRepository;
 
   @Autowired
-  public FileResourceServiceImpl(
+  public FileResourceBinaryServiceImpl(
       FileResourceMetadataRepository metadataRepository,
       FileResourceBinaryRepository binaryRepository) {
-    super(metadataRepository);
     this.binaryRepository = binaryRepository;
-  }
-
-  @Override
-  public FileResource createByMimeType(MimeType mimeType) {
-    return ((FileResourceMetadataRepository) repository).createByMimeType(mimeType);
-  }
-
-  @Override
-  public FileResource get(UUID uuid) {
-    return repository.findOne(uuid);
+    this.metadataRepository = metadataRepository;
   }
 
   @Override
@@ -46,7 +33,7 @@ public class FileResourceServiceImpl extends IdentifiableServiceImpl<FileResourc
       throws IdentifiableServiceException {
     try {
       fileResource = binaryRepository.save(fileResource, binaryData);
-      fileResource = repository.save(fileResource);
+      fileResource = metadataRepository.save(fileResource);
       return fileResource;
     } catch (ResourceIOException e) {
       LOGGER.error("Cannot save fileResource " + fileResource.getFilename() + ": ", e);
