+++ dc-cudami-server/dc-cudami-server-webapp/src/main/java/de/digitalcollections/cudami/server/controller/identifiable/resource/FileResourceMetadataController.java
@@ -1,39 +1,26 @@
 package de.digitalcollections.cudami.server.controller.identifiable.resource;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import de.digitalcollections.cudami.server.business.api.service.LocaleService;
 import de.digitalcollections.cudami.server.business.api.service.exceptions.IdentifiableServiceException;
-import de.digitalcollections.cudami.server.business.api.service.identifiable.resource.FileResourceService;
+import de.digitalcollections.cudami.server.business.api.service.identifiable.resource.FileResourceMetadataService;
 import de.digitalcollections.model.api.identifiable.resource.FileResource;
-import de.digitalcollections.model.api.identifiable.resource.MimeType;
 import de.digitalcollections.model.api.paging.PageRequest;
 import de.digitalcollections.model.api.paging.PageResponse;
 import de.digitalcollections.model.api.paging.Sorting;
 import de.digitalcollections.model.api.paging.enums.Direction;
 import de.digitalcollections.model.api.paging.enums.NullHandling;
-import de.digitalcollections.model.impl.identifiable.parts.LocalizedTextImpl;
-import de.digitalcollections.model.impl.identifiable.parts.structuredcontent.LocalizedStructuredContentImpl;
 import de.digitalcollections.model.impl.paging.OrderImpl;
 import de.digitalcollections.model.impl.paging.PageRequestImpl;
 import de.digitalcollections.model.impl.paging.SortingImpl;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InvalidObjectException;
-import java.net.URLDecoder;
-import java.nio.charset.StandardCharsets;
 import java.util.Locale;
 import java.util.Objects;
 import java.util.UUID;
-import javax.servlet.http.HttpServletRequest;
-import org.apache.tomcat.util.http.fileupload.FileItemIterator;
-import org.apache.tomcat.util.http.fileupload.FileItemStream;
-import org.apache.tomcat.util.http.fileupload.FileUploadException;
-import org.apache.tomcat.util.http.fileupload.servlet.ServletFileUpload;
 import org.jsondoc.core.annotation.Api;
 import org.jsondoc.core.annotation.ApiMethod;
 import org.jsondoc.core.annotation.ApiPathParam;
 import org.jsondoc.core.annotation.ApiQueryParam;
 import org.jsondoc.core.annotation.ApiResponseObject;
+import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -46,23 +33,18 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 @RestController
 @Api(description = "The fileresource controller", name = "Fileresource controller")
-public class FileResourceController {
+public class FileResourceMetadataController {
 
-  private static final org.slf4j.Logger LOGGER =
-      LoggerFactory.getLogger(FileResourceController.class);
+  private static final Logger LOGGER =
+      LoggerFactory.getLogger(FileResourceMetadataController.class);
 
-  @Autowired FileResourceService fileResourceService;
+  @Autowired FileResourceMetadataService fileResourceService;
 
   @Autowired LocaleService localeService;
-
-  @Autowired private ObjectMapper objectMapper;
 
   @ApiMethod(description = "Get all fileresources")
   @GetMapping(
@@ -130,59 +112,14 @@ public class FileResourceController {
     return new ResponseEntity<>(fileResource, HttpStatus.OK);
   }
 
-  // FIXME solve conflict of similar post with upload or pure metadata
-  @ApiMethod(description = "Save a newly created fileresource")
-  @PostMapping(
-      value = {"/latest/fileresources2", "/v2/fileresources2"},
-      produces = "application/json")
-  @ApiResponseObject
-  public FileResource save(@RequestBody FileResource fileResource)
-      throws IdentifiableServiceException {
-    return fileResourceService.save(fileResource);
-  }
-
   @ApiMethod(description = "Save a newly created fileresource")
   @PostMapping(
       value = {"/latest/fileresources", "/v2/fileresources"},
       produces = "application/json")
   @ApiResponseObject
-  public ResponseEntity<FileResource> save(
-      @RequestParam("fileresource") String resourceJson,
-      @RequestPart("binaryData") MultipartFile file,
-      RedirectAttributes redirectAttributes,
-      HttpServletRequest request) {
-    FileResource fileResource;
-    try {
-      // FIXME: is it really necessary to handle string and convert to object (no direct support of
-      // spring boot/mvc?)
-      fileResource = objectMapper.readValue(resourceJson, FileResource.class);
-      LOGGER.info("resource: " + fileResource.getLabel().getText());
-    } catch (IOException ex) {
-      LOGGER.error(
-          "Error: Cannot convert resource json to resource or read file bytes: '"
-              + resourceJson
-              + "'",
-          ex);
-      return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-    }
-
-    InputStream inputStream = null;
-    if (!file.isEmpty()) {
-      try {
-        inputStream = file.getInputStream();
-      } catch (IOException ex) {
-        LOGGER.error("Error getting binary data from uploaded file", ex);
-        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-      }
-    }
-
-    try {
-      fileResourceService.save(fileResource, inputStream);
-    } catch (IdentifiableServiceException ex) {
-      LOGGER.error("Error saving fileresource and binary data for uploaded file", ex);
-      return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-    return new ResponseEntity<>(fileResource, HttpStatus.OK);
+  public FileResource save(@RequestBody FileResource fileResource)
+      throws IdentifiableServiceException {
+    return fileResourceService.save(fileResource);
   }
 
   @ApiMethod(description = "Update a fileresource")
@@ -195,60 +132,5 @@ public class FileResourceController {
       throws IdentifiableServiceException {
     assert Objects.equals(uuid, fileResource.getUuid());
     return fileResourceService.update(fileResource);
-  }
-
-  @PostMapping(value = {"/latest/fileresources/new/upload", "/v2/fileresources/new/upload"})
-  @ApiResponseObject
-  public FileResource apiUpload(HttpServletRequest request) throws IOException {
-    FileResource fileResource = null;
-
-    InputStream stream = null;
-    try {
-      boolean isMultipart = ServletFileUpload.isMultipartContent(request);
-      if (!isMultipart) {
-        throw new InvalidObjectException("no multipart content");
-      }
-
-      ServletFileUpload upload = new ServletFileUpload();
-
-      FileItemIterator iter = upload.getItemIterator(request);
-      while (iter.hasNext()) {
-        FileItemStream item = iter.next();
-        if (!item.isFormField()) {
-          stream = item.openStream();
-          String originalFilename = item.getName();
-          originalFilename = URLDecoder.decode(originalFilename, StandardCharsets.UTF_8.toString());
-          String contentType = item.getContentType();
-
-          fileResource = fileResourceService.createByMimeType(MimeType.fromTypename(contentType));
-          fileResource.setFilename(originalFilename);
-          LOGGER.info("filename = " + fileResource.getFilename());
-
-          // set label to originalfilename for now. can be changed in next step of user input
-          fileResource.setDescription(new LocalizedStructuredContentImpl());
-          fileResource.setLabel(
-              new LocalizedTextImpl(
-                  new Locale(localeService.getDefaultLanguage()), originalFilename));
-
-          fileResource = fileResourceService.save(fileResource, stream);
-          LOGGER.info(
-              "saved file '"
-                  + fileResource.getUri().toString()
-                  + "' ("
-                  + fileResource.getSizeInBytes()
-                  + " bytes)");
-
-          stream.close();
-        }
-      }
-    } catch (FileUploadException | IOException | IdentifiableServiceException ex) {
-      LOGGER.error("Error getting binary data from uploaded file", ex);
-    } finally {
-      if (stream != null) {
-        stream.close();
-      }
-    }
-
-    return fileResource;
   }
 }
